requirements.txt
matplotlib==3.3.2
numpy==1.19.2
six==1.15.0
tensorflow==2.3.0
object_detection==0.1
opencv_python==4.4.0.46
scipy==1.5.2
ipython==7.19.0
Pillow==8.1.0
pyzmq==21.0.1
